pipeline/requirements.txt
torch
transformers
pytesseract
Pillow
ollama
IndicTransToolkit @ git+https://github.com/VarunGumma/IndicTransToolkit